[TRL-50] [feat] logincheck client component 분리
